+++ xmake/core/sandbox/modules/import/core/project/config.lua
@@ -116,12 +116,7 @@ function sandbox_core_project_config.check()
     -- get the current platform 
     local instance, errors = platform.load()
     if instance then
-
-        -- do check
-        local on_check = instance:script("config_check")
-        if on_check then
-            on_check(instance)
-        end
+        instance:check()
     else
         raise(errors)
     end
