+++ xmake/platforms/macosx/xmake.lua
@@ -36,11 +36,8 @@ platform("macosx")
     -- set install directory
     set_installdir("/usr/local")
 
-    -- on check project configuration
---    on_config_check("config")
-
-    -- on check global configuration
-    on_global_check("global")
+    -- on check 
+    on_check("check")
 
     -- on load
     on_load("load")
