+++ xmake/platforms/mingw/xmake.lua
@@ -33,8 +33,8 @@ platform("mingw")
     -- set formats
     set_formats {static = "$(name).lib", object = "$(name).obj", shared = "$(name).dll", binary = "$(name).exe", symbol = "$(name).pdb"}
 
-    -- on check project configuration
-    on_config_check("config")
+    -- on check
+    on_check("check")
 
     -- on load
     on_load("load")
