+++ xmake/platforms/msys/check.lua
@@ -15,29 +15,53 @@
 -- Copyright (C) 2015-2020, TBOOX Open Source Group.
 --
 -- @author      ruki
--- @file        config.lua
+-- @file        check.lua
 --
 
 -- imports
 import("core.project.config")
 import("core.base.singleton")
-import("core.platform.environment")
+import("detect.sdks.find_cross_toolchain")
 import("private.platform.toolchain")
 import("private.platform.check_arch")
-import("private.platform.check_vstudio")
 import("private.platform.check_toolchain")
+
+-- check the architecture
+function _check_arch()
+
+    -- get the architecture
+    local arch = config.get("arch")
+    if not arch then
+
+        -- init the default architecture
+        config.set("arch", config.get("cross") and "none" or os.subarch())
+
+        -- trace
+        print("checking for the architecture ... %s", config.get("arch"))
+    end
+end
 
 -- get toolchains
 function _toolchains()
 
+    -- find cross toolchain
+    local cross = ""
+    local cross_toolchain = find_cross_toolchain(config.get("sdk") or config.get("bin"), {bindir = config.get("bin"), cross = config.get("cross")})
+    if cross_toolchain then
+        config.set("cross", cross_toolchain.cross, {readonly = true, force = true})
+        config.set("bin", cross_toolchain.bindir, {readonly = true, force = true})
+        cross = cross_toolchain.cross
+    end
+
     -- init toolchains
     local cc         = toolchain("the c compiler")
     local cxx        = toolchain("the c++ compiler")
-    local mrc        = toolchain("the resource compiler")
     local ld         = toolchain("the linker")
     local sh         = toolchain("the shared library linker")
     local ar         = toolchain("the static library archiver")
     local ex         = toolchain("the static library extractor")
+    local mm         = toolchain("the objc compiler")
+    local mxx        = toolchain("the objc++ compiler")
     local as         = toolchain("the assember")
     local gc         = toolchain("the golang compiler")
     local gc_ld      = toolchain("the golang linker")
@@ -52,40 +76,56 @@ function _toolchains()
     local rc_ar      = toolchain("the rust static library archiver")
     local cu         = toolchain("the cuda compiler")
     local cu_ld      = toolchain("the cuda linker")
-    local toolchains = {cc = cc, cxx = cxx, mrc = mrc, as = as, ld = ld, sh = sh, ar = ar, ex = ex, 
+    local cu_ccbin   = toolchain("the cuda host c++ compiler")
+    local toolchains = {cc = cc, cxx = cxx, as = as, ld = ld, sh = sh, ar = ar, ex = ex, 
+                        mm = mm, mxx = mxx,
                         gc = gc, ["gcld"] = gc_ld, ["gcar"] = gc_ar,
                         dc = dc, ["dcld"] = dc_ld, ["dcsh"] = dc_sh, ["dcar"] = dc_ar,
                         rc = rc, ["rcld"] = rc_ld, ["rcsh"] = rc_sh, ["rcar"] = rc_ar,
-                        cu = cu, ["cu-ld"] = cu_ld}
+                        cu = cu, ["cu-ld"] = cu_ld, ["cu-ccbin"] = cu_ccbin}
 
     -- init the c compiler
-    cc:add("cl.exe")
+    cc:add("$(env CC)", {name = "gcc", cross = cross}, {name = "clang", cross = cross})
 
     -- init the c++ compiler
-    cxx:add("cl.exe")
-
-    -- init the resource compiler
-    mrc:add("rc.exe")
+    cxx:add("$(env CXX)")
+    cxx:add({name = "gcc", cross = cross})
+    cxx:add({name = "clang", cross = cross})
+    cxx:add({name = "g++", cross = cross})
+    cxx:add({name = "clang++", cross = cross})
 
     -- init the assember
-    local arch = config.get("arch")
-    if arch and arch:find("64") then
-        as:add("ml64.exe")
-    else
-        as:add("ml.exe")
-    end
+    as:add("$(env AS)", {name = "gcc", cross = cross}, {name = "clang", cross = cross})
 
     -- init the linker
-    ld:add("link.exe")
+    ld:add("$(env LD)", "$(env CXX)")
+    ld:add({name = "g++", cross = cross})
+    ld:add({name = "gcc", cross = cross})
+    ld:add({name = "clang++", cross = cross})
+    ld:add({name = "clang", cross = cross})
 
     -- init the shared library linker
-    sh:add("link.exe -dll")
+    sh:add("$(env SH)", "$(env CXX)")
+    sh:add({name = "g++", cross = cross})
+    sh:add({name = "gcc", cross = cross})
+    sh:add({name = "clang++", cross = cross})
+    sh:add({name = "clang", cross = cross})
 
     -- init the static library archiver
-    ar:add("link.exe -lib")
+    ar:add("$(env AR)", {name = "ar", cross = cross})
 
     -- init the static library extractor
-    ex:add("lib.exe")
+    ex:add("$(env AR)", {name = "ar", cross = cross})
+
+    -- init the objc compiler
+    mm:add("$(env MM)", {name = "clang", cross = cross}, {name = "gcc", cross = cross})
+
+    -- init the objc++ compiler
+    mxx:add("$(env MXX)")
+    mxx:add({name = "clang", cross = cross})
+    mxx:add({name = "clang++", cross = cross})
+    mxx:add({name = "gcc", cross = cross})
+    mxx:add({name = "g++", cross = cross})
 
     -- init the golang compiler and linker
     gc:add("$(env GC)", "go", "gccgo")
@@ -105,8 +145,11 @@ function _toolchains()
     rc_ar:add("$(env RC)", "rustc")
 
     -- init the cuda compiler and linker
-    cu:add("nvcc", "clang")
+    cu:add("nvcc", "clang++", "clang")
     cu_ld:add("nvcc")
+    if not cross or cross == "" then
+        cu_ccbin:add("$(env CXX)", "$(env CC)", "gcc", "clang", "g++", "clang++")
+    end
 
     return toolchains
 end
@@ -116,24 +159,14 @@ function main(platform, name)
 
     -- only check the given config name?
     if name then
-        local toolchain = singleton.get("windows.toolchains." .. (config.get("arch") or os.arch()), _toolchains)[name]
+        local toolchain = singleton.get("msys.toolchains", _toolchains)[name]
         if toolchain then
-            environment.enter("toolchains")
             check_toolchain(config, name, toolchain)
-            environment.leave("toolchains")
         end
     else
 
-        -- check arch 
-        check_arch(config)
-
-        -- check vstudio
-        local cc  = path.basename(config.get("cc") or "cl"):lower()
-        local cxx = path.basename(config.get("cxx") or "cl"):lower()
-        local mrc = path.basename(config.get("mrc") or "rc"):lower()
-        if cc == "cl" or cxx == "cl" or mrc == "rc" then
-            check_vstudio(config)
-        end
+        -- check arch
+        _check_arch()
     end
 end
