+++ xmake/platforms/watchos/xmake.lua
@@ -33,11 +33,8 @@ platform("watchos")
     -- set formats
     set_formats {static = "lib$(name).a", object = "$(name).o", shared = "lib$(name).dylib", symbol = "$(name).dSYM"}
 
-    -- on check project configuration
-    on_config_check("config")
-
-    -- on check global configuration
-    on_global_check("global")
+    -- on check 
+    on_check("check")
 
     -- on load
     on_load("load")
